Beginning Work On Wall Run Functions.
Added OnActorHit event to initiate required checks for beginning a wall run after the character hits an another actor. Added IsCustomMovementMode function for ease of comparing the UCharacterMovementComponent's EMovementMode with UParkourMovementComponent's ECustomMovementMode. Added check wall run which will run through all of the sub-checks required for wall running, and will then begin the wall run. Added CheckWallRunFloor to make sure that the player has not yet hit the ground while running. Added functions for finding the wall run end vectors for both the left and the right side of the character. Added IsValidWallRunVector to validate if the traces done using the wall run end vectors found it suitable to begin a wall run. Added PlayerToWallVector to which will be used to attach the player to the wall.